@@ -17,40 +17,40 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+2
 
-analyzer:
-  enable-experiment:
-  - non-nullable
+# analyzer:
+#   enable-experiment:
+#   - non-nullable
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.5 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   avatar_view: ^1.0.0+4
-  badges: ^2.0.1
+  badges: ^3.1.1
   cached_network_image: ^3.0.0
   comment_box: ^0.0.16
   cupertino_icons: ^1.0.2
-  dynamic_overflow_menu_bar: ^0.1.2
-  flutter_lints: ^1.0.3
+  # dynamic_overflow_menu_bar: ^0.1.2
+  # flutter_lints: ^2.0.2
   get: ^4.1.2
-  google_fonts: ^2.1.0
+  google_fonts: ^5.1.0
   google_nav_bar: ^5.0.5
-  http: ^0.13.1
-  image_picker: ^0.8.0+3
+  http: ^1.1.0
+  image_picker: ^1.0.1
   line_icons: ^2.0.1
-  material_design_icons_flutter: ^4.0.5345
+  material_design_icons_flutter: ^7.0.7296
   material_floating_search_bar: ^0.3.4
   path_provider: ^2.0.2
-  permission_handler: ^8.1.1
+  permission_handler: ^10.4.3
   shared_preferences: ^2.0.5
   simple_speed_dial: ^0.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_svg: ^0.22.0
-  picture_in_picture: ^0.0.4
+  flutter_svg: ^2.0.7
+  # picture_in_picture: ^0.0.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
