@@ -4,15 +4,14 @@ import 'package:flutter_facebook_responsive_ui/models/models.dart';
 import 'package:flutter_facebook_responsive_ui/profile/profile_avatar.dart';
 import 'package:flutter_facebook_responsive_ui/widgets/post_widgets/more_options_post_widget.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'more_options_post_widget.dart';
 
 class TitlePostContainer extends StatelessWidget {
   final Post post;
   final ind;
 
   const TitlePostContainer({
-    Key key,
-    @required this.post,
+    Key? key,
+    required this.post,
     this.ind,
   }) : super(key: key);
 
@@ -68,8 +67,8 @@ class _PostHeader extends StatelessWidget {
   final Post post;
 
   const _PostHeader({
-    Key key,
-    @required this.post,
+    Key? key,
+    required this.post,
   }) : super(key: key);
 
   @override
@@ -119,7 +118,7 @@ class _PostHeader extends StatelessWidget {
           icon: const Icon(Icons.more_horiz),
           onPressed: () async {
             SharedPreferences prefs = await SharedPreferences.getInstance();
-            String isDiary = prefs.getString("diary");
+            String? isDiary = prefs.getString("diary");
             print("diary ==== " + isDiary.toString());
             if (isDiary == "yes") {
               print("3456789");
@@ -136,163 +135,17 @@ class _PostHeader extends StatelessWidget {
 
 //SHOW MORE OPTIONS DIALOG BOX
 
-//things like (Number of Likes,comments,shares etc)
-// class _PostStats extends StatefulWidget {
-//   final Post post;
-//   final ind;
-//   const _PostStats({
-//     Key key,
-//     @required this.post,
-//     this.ind,
-//   }) : super(key: key);
-
-//   @override
-//   State<_PostStats> createState() => _PostStatsState();
-// }
-
-// class _PostStatsState extends State<_PostStats> {
-//   bool IsLiked = false;
-//   @override
-//   void initState() {
-//     super.initState();
-//   }
-
-//   fecthData() async {
-//     SharedPreferences prefs = await SharedPreferences.getInstance();
-//     var userId = prefs.getString(
-//       "userId",
-//     );
-
-//     print(userId);
-
-//     /// api calling
-//     // var res = await http.get(Uri.parse(
-//     //     API + "/Post/PostLiked?userID=$userId&postId=${widget.post.pid}"));
-//     // print(res.body);
-//     // print(Get.find<PostController1>().session_posts[widget.ind].likes);
-//     // print(Get.find<PostController1>().session_posts[widget.ind].isAlreadyLiked);
-
-//     // Get.find<PostController1>().updatePostsLikedSessionwall(widget.ind);
-
-//     setState(() {});
-
-//     // setState(() {
-//     //   widget.post.isAlreadyLiked
-//     //       ? widget.post.isAlreadyLiked = false
-//     //       : widget.post.isAlreadyLiked = true;
-//     // });
-//   }
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return GetBuilder<PostController1>(
-//       init: PostController1(),
-//       initState: (_) {},
-//       builder: (_) {
-//         return Column(
-//           children: [
-//             Row(
-//               children: [
-//                 Container(
-//                   padding: const EdgeInsets.all(4.0),
-//                   decoration: BoxDecoration(
-//                     color: Palette.facebookBlue,
-//                     shape: BoxShape.circle,
-//                   ),
-//                   child: const Icon(
-//                     Icons.thumb_up,
-//                     size: 10.0,
-//                     color: Colors.white,
-//                   ),
-//                 ),
-//                 const SizedBox(width: 4.0),
-//                 Expanded(
-//                   //how to make it dynamic for all walls?
-//                   child: Text(
-//                     '${_.session_posts[widget.ind].likes}',
-//                     style: TextStyle(
-//                       color: Colors.grey[600],
-//                     ),
-//                   ),
-//                 ),
-//                 // Text(
-//                 //   '${post.comments} Comments',
-//                 //   style: TextStyle(
-//                 //     color: Colors.grey[600],
-//                 //   ),
-//                 // ),
-//                 //const SizedBox(width: 8.0),
-//                 //this code showed how many shares for this post
-//                 // Text(
-//                 //   '${post.shares} Shares',
-//                 //   style: TextStyle(
-//                 //     color: Colors.grey[600],
-//                 //   ),
-//                 // )
-//               ],
-//             ),
-//             const Divider(),
-//             //Buttons of the page (like,comment,share)
-//             Row(
-//               children: [
-//                 _PostButton(
-//                     icon: Icon(
-//                       _.session_posts[widget.ind].isAlreadyLiked
-//                           ? MdiIcons.thumbUp
-//                           : MdiIcons.thumbUpOutline,
-//                       color: Colors.blue[600],
-//                       size: 20.0,
-//                     ),
-//                     label: 'Like',
-//                     onTap: () {
-//                       fecthData();
-
-//                       ;
-//                     }),
-//                 _PostButton(
-//                     icon: Icon(
-//                       MdiIcons.commentOutline,
-//                       color: Colors.grey[600],
-//                       size: 20.0,
-//                     ),
-//                     label: 'Comment',
-//                     onTap: () {
-//                       // Get.to(TestMe());
-//                     }),
-//                 _PostButton(
-//                   icon: Icon(
-//                     MdiIcons.shareOutline,
-//                     color: Colors.grey[600],
-//                     size: 25.0,
-//                   ),
-//                   label: 'Share',
-//                   onTap: () {
-//                     Get.to(Share(pid: widget.post));
-//                   },
-//                 )
-//               ],
-//             ),
-//             const Divider(
-//               color: Colors.black,
-//             ),
-//           ],
-//         );
-//       },
-//     );
-//   }
-// }
-
 //button code like,comment
 class _PostButton extends StatelessWidget {
   final Icon icon;
   final String label;
-  final Function onTap;
+  final void Function() onTap;
 
   const _PostButton({
-    Key key,
-    @required this.icon,
-    @required this.label,
-    @required this.onTap,
+    Key? key,
+    required this.icon,
+    required this.label,
+    required this.onTap,
   }) : super(key: key);
 
   @override
