@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_facebook_responsive_ui/config/api.dart';
 import 'package:flutter_facebook_responsive_ui/controllers/post_controller.dart';
 import 'package:flutter_facebook_responsive_ui/models/models.dart';
@@ -96,7 +97,6 @@ class _GroupPageState extends State<GroupPage> {
             SliverAppBar(
               stretch: true,
               elevation: 5.0,
-              brightness: Brightness.light,
               backgroundColor: Colors.green[400]!.withOpacity(.79),
               flexibleSpace: FlexibleSpaceBar(
                 // background: Image.asset(
@@ -133,6 +133,7 @@ class _GroupPageState extends State<GroupPage> {
                 //       // Get.changeTheme(ThemeData.dark());
                 //     }),
               ],
+              systemOverlayStyle: SystemUiOverlayStyle.dark,
             ),
             //   SliverToBoxAdapter(child: ForAI()),
             GetBuilder<PostController1>(
